+++ packages/observable-client/src/chainHead/chainHead.ts
@@ -28,6 +28,7 @@ import {
   shareReplay,
   switchMap,
   take,
+  takeWhile,
   tap,
 } from "rxjs"
 
@@ -50,7 +51,7 @@ import type {
   RuntimeContext,
   SystemEvent,
 } from "./streams"
-import { getFollow$, getPinnedBlocks$ } from "./streams"
+import { getFollow$, getPinnedBlocks$, toBlockInfo } from "./streams"
 import { getTrackTx } from "./track-tx"
 import { getValidateTx } from "./validate-tx"
 
@@ -66,20 +67,15 @@ export type BlockInfo = {
   hash: string
   number: number
   parent: string
+  hasNewRuntime: boolean
 }
-
-const toBlockInfo = ({ hash, number, parent }: PinnedBlock): BlockInfo => ({
-  hash,
-  number,
-  parent,
-})
 
 export const getChainHead$ = (
   chainHead: ChainHead,
   getCachedMetadata: (codeHash: string) => Observable<Uint8Array | null>,
   setCachedMetadata: (codeHash: string, metadataRaw: Uint8Array) => void,
 ) => {
-  const { getFollower, startFollow, follow$, getHeader } = getFollow$(chainHead)
+  const { getFollower, unfollow, follow$, getHeader } = getFollow$(chainHead)
   const lazyFollower = withLazyFollower(getFollower)
   const { withRecovery, withRecoveryFn } = getWithRecovery()
 
@@ -155,7 +151,7 @@ export const getChainHead$ = (
     // ":code" => "0x3a636f6465"
     stg(blockHash, "hash", "0x3a636f6465", null).pipe(map((x) => x!))
 
-  const newBlocks$ = new Subject<BlockInfo>()
+  const newBlocks$ = new Subject<BlockInfo | null>()
   const pinnedBlocks$ = getPinnedBlocks$(
     follow$,
     withRefcount(withRecoveryFn(fromAbortControllerFn(lazyFollower("call")))),
@@ -411,23 +407,12 @@ export const getChainHead$ = (
     }),
     shareReplay(1),
   )
-
   // calling `unfollow` also kills the subscription due to the fact
   // that `follow$` completes, which makes all other streams to
   // also complete (or error, in the case of ongoing operations)
   merge(runtime$, bestBlocks$).subscribe({
     error() {},
   })
-
-  let unfollow = noop
-  let started: boolean | null = false
-  let nSubscribers: number = 0
-  const start = (_nSubscribers: number) => {
-    nSubscribers += _nSubscribers
-    started = true
-
-    unfollow = startFollow()
-  }
 
   const getRuntime$ = (codeHash: string): Observable<RuntimeContext | null> =>
     merge(
@@ -444,7 +429,7 @@ export const getChainHead$ = (
     let hash = blockHash || "finalized"
     hash = pinnedBlocks$.state[hash as "best" | "finalized"] || hash
 
-    if (!started || !pinnedBlocks$.state.blocks.has(hash)) {
+    if (!pinnedBlocks$.state.blocks.has(hash)) {
       if (shouldThrow) throw new BlockNotPinnedError(hash, "holdBlock")
       return noop
     }
@@ -481,45 +466,34 @@ export const getChainHead$ = (
         return subscription
       })
 
-  return [
-    {
-      follow$,
-      finalized$,
-      best$,
-      bestBlocks$,
-      newBlocks$,
-      runtime$,
-      metadata$,
-      genesis$,
-      getRuntime$,
+  return {
+    follow$,
+    unfollow,
+    finalized$,
+    best$,
+    bestBlocks$,
+    newBlocks$: newBlocks$.pipe(takeWhile(Boolean)),
+    runtime$,
+    metadata$,
+    genesis$,
+    getRuntime$,
 
-      header$,
-      body$,
-      call$,
-      storage$,
-      storageQueries$,
-      eventsAt$,
+    header$,
+    body$,
+    call$,
+    storage$,
+    storageQueries$,
+    eventsAt$,
 
-      withHodl,
-      holdBlock,
-      trackTx$,
-      trackTxWithoutEvents$,
-      validateTx$,
-      pinnedBlocks$,
-      withRuntime,
-      getRuntimeContext$: withOptionalHash$(getRuntimeContext$),
-      unfollow: () => {
-        if (started == null) return
-        nSubscribers--
-        if (started && !nSubscribers) {
-          started = null
-          unfollow()
-          unfollow = noop
-        }
-      },
-    },
-    start,
-  ] as const
+    withHodl,
+    holdBlock,
+    trackTx$,
+    trackTxWithoutEvents$,
+    validateTx$,
+    pinnedBlocks$,
+    withRuntime,
+    getRuntimeContext$: withOptionalHash$(getRuntimeContext$),
+  }
 }
 
-export type ChainHead$ = ReturnType<typeof getChainHead$>[0]
+export type ChainHead$ = ReturnType<typeof getChainHead$>