+++ packages/client/CHANGELOG.md
@@ -11,6 +11,8 @@
 
 - Update to @noble v2 libraries. This bumps the NodeJS required version to 20.19
 - `blocks$` completes when there is no block continuity.
+- **Smoldot:**
+  - Update smoldot to `2.0.37`
 
 ## 1.16.4 - 2025-09-01