+++ packages/client/CHANGELOG.md
@@ -13,7 +13,7 @@
 - Update to @noble v2 libraries. This bumps the NodeJS required version to 20.19
 - `blocks$` completes when there is no block continuity.
 - **Smoldot:**
-  - Update smoldot to `2.0.37`
+  - Update smoldot to `2.0.38`
 - **Chains:**:
   - Update `lightSyncState`